Fixed acceptance test for start time / end time slider range.
It turned out to be that when you run the tests in
cms/djangoapps/contentstore/features/video.feature
all at once, then test # 11 starts with transcripts already turned on. When you
run just the test # 11, then transcripts are not available (from previous tests).
That's why to make width consistent, we have to make sure that transcripts are
hidden.